+++ packages/mountainsort2/src/p_detect_events.cpp
@@ -12,21 +12,21 @@ QVector<double> subsample_events(const QVector<double>& X, double subsample_fact
 bool p_detect_events(QString timeseries, QString event_times_out, P_detect_events_opts opts)
 {
     DiskReadMda32 X(timeseries);
-    int M = X.N1();
-    int N = X.N2();
+    bigint M = X.N1();
+    bigint N = X.N2();
 
     printf("Collecting data vector...\n");
     QVector<double> data(N);
     if (opts.central_channel > 0) {
-        for (int i = 0; i < N; i++) {
+        for (bigint i = 0; i < N; i++) {
             data[i] = X.value(opts.central_channel - 1, i);
         }
     }
     else {
-        for (int i = 0; i < N; i++) {
+        for (bigint i = 0; i < N; i++) {
             double best_value = 0;
-            int best_m = 0;
-            for (int m = 0; m < M; m++) {
+            bigint best_m = 0;
+            for (bigint m = 0; m < M; m++) {
                 double val = X.value(m, i);
                 if (opts.sign < 0)
                     val = -val;
@@ -53,7 +53,7 @@ bool p_detect_events(QString timeseries, QString event_times_out, P_detect_event
 
     printf("Creating result array...\n");
     Mda ret(1, event_times.count());
-    for (int j = 0; j < event_times.count(); j++) {
+    for (bigint j = 0; j < event_times.count(); j++) {
         ret.setValue(event_times[j], j);
     }
     printf("Writing result...\n");
@@ -66,12 +66,12 @@ QVector<double> detect_events(const QVector<double>& X, double detect_threshold,
     double stdev = MLCompute::stdev(X);
     double threshold2 = detect_threshold * stdev;
 
-    int N = X.count();
-    QVector<int> to_use(N);
+    bigint N = X.count();
+    QVector<bigint> to_use(N);
     to_use.fill(0);
-    int last_best_ind = 0;
+    bigint last_best_ind = 0;
     double last_best_val = 0;
-    for (int n = 0; n < N; n++) {
+    for (bigint n = 0; n < N; n++) {
         double val = X[n];
         if (sign < 0)
             val = -val;
@@ -96,7 +96,7 @@ QVector<double> detect_events(const QVector<double>& X, double detect_threshold,
         }
     }
     QVector<double> times;
-    for (int n = 0; n < N; n++) {
+    for (bigint n = 0; n < N; n++) {
         if (to_use[n]) {
             times << n;
         }
@@ -107,14 +107,14 @@ QVector<double> detect_events(const QVector<double>& X, double detect_threshold,
 double pseudorandomnumber(double i)
 {
     double ret = sin(i + cos(i));
-    ret = (ret + 5) - (int)(ret + 5);
+    ret = (ret + 5) - (bigint)(ret + 5);
     return ret;
 }
 
 QVector<double> subsample_events(const QVector<double>& X, double subsample_factor)
 {
     QVector<double> ret;
-    for (int i = 0; i < X.count(); i++) {
+    for (bigint i = 0; i < X.count(); i++) {
         double randnum = pseudorandomnumber(i);
         if (randnum <= subsample_factor)
             ret << X[i];
